pyproject.toml
[tool.poetry]
authors = ["Paulo S. Costa <Paulo.S.Costa5@gmail.com>"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
description = "nbpreview"
documentation = "https://nbpreview.readthedocs.io"
homepage = "https://github.com/paw-lu/nbpreview"
license = "MIT"
name = "nbpreview"
readme = "README.md"
repository = "https://github.com/paw-lu/nbpreview"
version = "0.9.1"

[tool.poetry.urls]
Changelog = "https://github.com/paw-lu/nbpreview/releases"
Issues = "https://github.com/paw-lu/nbpreview/issues"

# dependencies-start

[tool.poetry.dependencies]
Jinja2 = ">=3.1.6"
Pillow = ">=8.3.1,<11.0.0"
Pygments = ">=2.19.2"
click-help-colors = ">=0.9.4"
html2text = ">=2025.4.15"
httpx = "^0.28.1"
ipython = "^9.6.0"
lxml = ">=6.0.2"
markdown-it-py = ">=3.0,<4.0"
mdit-py-plugins = ">=0.5.0"
nbformat = { extras = ["fast"], version = ">=5.10.4" }
picharsso = ">=2.0.1"
pylatexenc = ">=2.10"
python = "^3.11.0"
rich = ">=14.2.0"
term-image = ">=0.7.2"
typer = "^0.19.2"
types-click = ">=7.1.8"
validators = "^0.35.0"
yarl = ">=1.22.0"

[tool.poetry.group.dev.dependencies]
Pygments = ">=2.19.2"
autoflake = ">=2.3.1"
black = { extras = ["jupyter"], version = ">=25.9.0" }
click = ">=8.3.0"
coverage = { extras = ["toml"], version = ">=7.10.7" }
darglint = ">=1.8.1"
flake8 = ">=7.3.0"
flake8-bandit = ">=4.1.1"
flake8-bugbear = ">=24.12.12"
flake8-docstrings = ">=1.7.0"
flake8-rst-docstrings = ">=0.4.0"
furo = ">=2025.9.25"
interrogate = ">=1.7.0"
ipykernel = ">=7.0.0"
isort = ">=7.0.0"
mypy = ">=1.18.2"
myst-parser = ">=4.0.1"
nbqa = ">=1.9.1"
pdbpp = ">=0.11.7"
pep8-naming = ">=0.15.1"
pre-commit = ">=4.3.0"
pre-commit-hooks = ">=6.0.0"
pytest = ">=8.4.2"
pytest-mock = ">=3.15.1"
pyupgrade = ">=3.21.0"
safety = ">=3.6.2"
sphinx = ">=8.2.3"
sphinx-autobuild = ">=2025.8.25"
sphinx-autodoc-typehints = ">=3.5.1"
sphinx-click = ">=6.1.0"
sphinx-copybutton = ">=0.5.2"
sphinx-design = ">=0.6.1"
sphinx-favicon = ">=1.0.1"
sphinxext-opengraph = ">=0.13.0"
tomli = ">=2.3.0"
typeguard = ">=4.4.4"
xdoctest = { extras = ["colors"], version = ">=1.3.0" }

# dependencies-end

[tool.poetry.scripts]
nbp = "nbpreview.__main__:app"
nbpreview = "nbpreview.__main__:app"

[tool.coverage.paths]
source = ["*/site-packages", "src"]
tests = ["*/tests", "tests"]

[tool.coverage.run]
branch = true
source = ["nbpreview", "tests"]

[tool.coverage.report]
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",
    # Typing overloads are never run
    "@typing.overload",
]
fail_under = 100
show_missing = true

[tool.mypy]
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict = true
warn_unreachable = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "click_help_colors,jsonschema,lxml.*,nbformat.*,PIL.*,picharsso.*,pygments,pylatexenc,pytest_mock,term_image,validators,tomli"

[[tool.mypy.overrides]]
module = "mdit_py_plugins.*"
no_implicit_reexport = false

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
markers = [
    "no_atty_mock: Test where sys.stdout.isatty should not be mocked",
    "no_typeguard: Tests typeguard should ignore",
]
norecursedirs = "tests/util"

[tool.interrogate]
color = true
fail-under = 100
verbose = 2

[tool.isort]
profile = "black"

[tool.nbqa.dont_skip_bad_cells]
black = true
flake8 = true
isort = true
pyupgrade = true

[tool.nbqa.addopts]
flake8 = ["--extend-ignore=D100"]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
